+++ lib/main.rb
@@ -3,5 +3,6 @@ require_relative 'web_crawler'
 scrape = Web_crawler.new 
 scrape.crawl
 
-scrape_whole = Whole_web.new 
-scrape_whole.crawl_whole
+# scrape_whole = Whole_web.new 
+# scrape_whole.crawl_whole
+
